Added pagescrollview lib
It's a lib of mine that I always use in my projects. It's very unknown, but it's actually really useful to have a fully working ScrollView out of the box.

I understand and won't be sad at all if this PR is rejected due to its low downloads number.

https://github.com/SrBrahma/pagescrollview

Thanks!